requirements.txt
fastapi
uvicorn
torch
torchvision
numpy
pillow
pymysql
peewee
aiofiles
pytest
httpx
piexif
seaborn
matplotlib
scikit-learn
python-dotenv